# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "xAI Grok API",
    "description": "API for xAI Grok services including image generation, image understanding, and chat",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/responses": {
      "post": {
        "tags": [
          "Responses API (Native Agentic Tools)"
        ],
        "summary": "Create response using xAI Responses API with native agentic tools",
        "description": "Create a response using xAI's new Responses API. This endpoint supports:\n    \n    - **Native Agentic Tools**: web_search, x_search, code_execution (executed by xAI's servers)\n    - **Server-side Storage**: Conversations stored for 30 days with automatic caching\n    - **Stateful Conversations**: Use previous_response_id to continue conversations\n    - **Citations**: Automatic citations from web and X searches\n    - **Mixed Tools**: Combine server-side tools with client-side function calling\n    \n    Set stream=true to receive a streaming response. Set XAI_NATIVE_TOOLS_ENABLED=true to enable server-side tools.",
        "operationId": "create_response_api_v1_responses_post",
        "responses": {
          "200": {
            "description": "Streaming response (when stream=true)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesResponse"
                }
              },
              "text/event-stream": {}
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/responses/{response_id}": {
      "get": {
        "tags": [
          "Responses API (Native Agentic Tools)"
        ],
        "summary": "Retrieve a previously stored response",
        "description": "Retrieve a response that was previously created with store=true.\n    \n    Responses are stored for 30 days on xAI's servers. Use this to:\n    - Retrieve conversation history\n    - Continue conversations without resending full history\n    - Access responses from different sessions\n    \n    The response_id is returned in the 'id' field when you create a response.",
        "operationId": "retrieve_response_api_v1_responses__response_id__get",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponsesResponse"
                }
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Responses API (Native Agentic Tools)"
        ],
        "summary": "Delete a previously stored response",
        "description": "Delete a response that was previously stored on xAI's servers.\n    \n    Use this to:\n    - Clean up conversation history\n    - Remove sensitive information\n    - Manage storage usage\n    \n    Once deleted, the response cannot be retrieved or used with previous_response_id.",
        "operationId": "delete_response_api_v1_responses__response_id__delete",
        "parameters": [
          {
            "name": "response_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Response Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Not Found"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Internal Server Error"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Citation": {
        "properties": {
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Title"
          },
          "snippet": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snippet"
          },
          "source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source",
            "description": "Source type: 'web' or 'x'"
          }
        },
        "type": "object",
        "title": "Citation",
        "description": "Citation information from searches"
      },
      "ErrorResponse": {
        "properties": {
          "error": {
            "type": "boolean",
            "title": "Error",
            "default": true
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ResponsesResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Response ID (can be used as previous_response_id)"
          },
          "object": {
            "type": "string",
            "title": "Object",
            "description": "Object type",
            "default": "response"
          },
          "created": {
            "type": "integer",
            "title": "Created",
            "description": "Unix timestamp"
          },
          "model": {
            "type": "string",
            "title": "Model",
            "description": "Model used"
          },
          "output": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Output",
            "description": "Array of output messages (flexible schema)"
          },
          "usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResponsesUsageMetrics"
              },
              {
                "type": "null"
              }
            ]
          },
          "server_side_tool_usage": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServerSideToolUsage"
              },
              {
                "type": "null"
              }
            ]
          },
          "citations": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/Citation"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Citations",
            "description": "Citations from web/X searches"
          },
          "finish_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finish Reason",
            "description": "Reason completion stopped"
          }
        },
        "additionalProperties": true,
        "type": "object",
        "required": [
          "id",
          "created",
          "model",
          "output"
        ],
        "title": "ResponsesResponse",
        "description": "Response model for xAI Responses API"
      },
      "ResponsesUsageMetrics": {
        "properties": {
          "input_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Input Tokens"
          },
          "output_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Output Tokens"
          },
          "total_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Total Tokens"
          },
          "reasoning_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reasoning Tokens"
          },
          "cached_tokens": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cached Tokens"
          }
        },
        "type": "object",
        "title": "ResponsesUsageMetrics",
        "description": "Usage metrics for Responses API"
      },
      "ServerSideToolUsage": {
        "properties": {
          "web_search_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Web Search Calls",
            "description": "Number of web search calls",
            "default": 0
          },
          "x_search_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "X Search Calls",
            "description": "Number of X search calls",
            "default": 0
          },
          "code_execution_calls": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Code Execution Calls",
            "description": "Number of code execution calls",
            "default": 0
          }
        },
        "type": "object",
        "title": "ServerSideToolUsage",
        "description": "Server-side tool usage metrics"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}